+++ handroll/composers/__init__.py
@@ -44,7 +44,12 @@ class Composers(object):
 
 
 class CopyComposer(Composer):
-    """Copy a source file to the destination"""
+    """Copy a source file to the destination.
+
+    ``CopyComposer`` is the default composer for any unrecognized file type.
+    The source file will be copied to the output directory unless there is a
+    file with an identical name and content already at the destination.
+    """
 
     def compose(self, template, source_file, out_dir):
         """Copy a file to the destination if the file does not exist or was
@@ -68,7 +73,29 @@ class CopyComposer(Composer):
 
 
 class MarkdownComposer(Composer):
-    """Compose HTML from Markdown"""
+    """Compose HTML from Markdown files (``.md``).
+
+    The first line of the file will be used as the ``title`` data for the
+    template. All following lines will be converted to HTML and sent to the
+    template as the ``content`` data.
+
+    The ``MarkdownComposer`` supports syntax highlighting using Pygments. Code
+    can be specified using "fenced code" triple backticks.
+
+    ::
+
+        ```python
+        class Foo(object):
+            '''This sample code would be highlighted in a Python style.'''
+        ```
+
+    Use ``pygmentize`` to create your desired CSS file. Refer to the
+    `Pygments documentation <http://pygments.org/docs/>`_ for more information.
+
+    .. code-block:: bash
+
+        $ pygmentize -S default -f html > pygments.css
+    """
 
     EXTENSIONS = [
         'codehilite',